+++ schemas/php.php
@@ -32,6 +32,9 @@ class Bank
     /** @var bool|null */
     public $CreditDocument;
 
+    /** @var bool **/
+    public $LegalCheque;
+
     /** @var string|null */
     public $SalaryPortability;