+++ schemas/python.py
@@ -70,6 +70,7 @@ class BankElement:
     pix_type: Optional[str]
     charge: bool
     creditDocument: bool
+    legalCheque: bool
     salaryPortability: str
     products: List[str]
     url: Optional[str]
@@ -78,7 +79,7 @@ class BankElement:
     date_registered: datetime
     date_updated: datetime
 
-    def __init__(self, compe: str, ispb: str, document: str, long_name: str, short_name: str, network: Optional[str], type_: Optional[str], pix_type: Optional[str], charge: bool, creditDocument: bool, salaryPortability: Optional[str], products: List[str], url: Optional[str], date_operation_started: Optional[str], date_pix_started: Optional[str], date_registered: datetime, date_updated: datetime) -> None:
+    def __init__(self, compe: str, ispb: str, document: str, long_name: str, short_name: str, network: Optional[str], type_: Optional[str], pix_type: Optional[str], charge: bool, creditDocument: bool, legalCheque: bool, salaryPortability: Optional[str], products: List[str], url: Optional[str], date_operation_started: Optional[str], date_pix_started: Optional[str], date_registered: datetime, date_updated: datetime) -> None:
         """Return the bank DTO from JSON file."""
         self.compe = compe
         self.ispb = ispb
@@ -90,6 +91,7 @@ class BankElement:
         self.pix_type = pix_type
         self.charge = charge
         self.creditDocument = creditDocument
+        self.legalCheque = legalCheque
         self.salaryPortability = salaryPortability
         self.products = products
         self.url = url
@@ -111,6 +113,7 @@ class BankElement:
         pix_type = from_union([from_none, from_str], obj.get("PixType"))
         charge = from_union([from_none, from_str], obj.get("Charge"))
         creditDocument = from_union([from_none, from_str], obj.get("CreditDocument"))
+        legalCheque = from_union([from_none, from_str], obj.get("LegalCheque"))
         salaryPortability = from_union([from_none, from_str], obj.get("SalaryPortability"))
         products = from_union([from_none, from_str], obj.get("Products"))
         url = from_union([from_none, from_str], obj.get("Url"))
@@ -118,7 +121,7 @@ class BankElement:
         date_pix_started = from_union([from_none, from_str], obj.get("DatePixStarted"))
         date_registered = from_datetime(obj.get("DateRegistered"))
         date_updated = from_datetime(obj.get("DateUpdated"))
-        return BankElement(compe, ispb, document, long_name, short_name, network, type_, pix_type, charge, creditDocument, salaryPortability, products, url, date_operation_started, date_pix_started, date_registered, date_updated)
+        return BankElement(compe, ispb, document, long_name, short_name, network, type_, pix_type, charge, creditDocument, legalCheque, salaryPortability, products, url, date_operation_started, date_pix_started, date_registered, date_updated)
 
     def to_dict(self) -> dict:
         result: dict = {}
@@ -132,6 +135,7 @@ class BankElement:
         result["PixType"] = from_union([from_none, from_str], self.pix_type)
         result["Charge"] =  from_union([from_none, from_str], self.charge)
         result["CreditDocument"] =  from_union([from_none, from_str], self.creditDocument)
+        result["LegalCheque"] = from_union([from_none, from_str], self.legalCheque)
         result["SalaryPortability"] =  from_union([from_none, from_str], self.salaryPortability)
         result["Products"] =  from_union([from_none, from_str], self.products)
         result["Url"] = from_union([from_none, from_str], self.url)